Revert "LPS-99657 Class that implements ConfigurationBeanDeclaration should be in package *.settings.definition"
This reverts commit 4cf79914fe90a961afdf33456c07fbb59b7ccb9b.
